Move url_to_local_dir to its own file
Sparse indexes use (almost) the same logic for calculating their
directory. However, the data stored there is in a different format.
Moving this function makes it easier to experiment with a sparse index
API without having to re-implement this logic.

In the long-term, I suspect we'll want to introduce a parallel type to
`Index` which exposes a subset of the API (probably just exposing
`crate_` and `index_config`). We may also want to offer a
`cached_crates` which would iterate over the already-downloaded crates,
and may want to offer different versions of `crate_` for "already
cached" vs "fetch all".
